requirements.txt
google-generativeai==0.8.5
Markdown==3.7
numpy==2.2.0
PyPDF2==3.0.1
sqlite-vec==0.1.6
bs4==4.13.4
